corrected api request bug
